Move `output_negative` into `tests/`
So it cannot be called from outside of tests!
