+++ core/src/libtx/build.rs
@@ -107,49 +107,6 @@ where
 	build_input(value, OutputFeatures::Coinbase, key_id)
 }
 
-/// Build a negative output. This function must not be used outside of tests.
-/// The commitment will be an inversion of the value passed in and the value is
-/// subtracted from the sum.
-pub fn output_negative<K, B>(value: u64, key_id: Identifier) -> Box<Append<K, B>>
-where
-	K: Keychain,
-	B: ProofBuild,
-{
-	Box::new(
-		move |build, acc| -> Result<(Transaction, BlindSum), Error> {
-			let (tx, sum) = acc?;
-
-			// TODO: proper support for different switch commitment schemes
-			let switch = SwitchCommitmentType::Regular;
-
-			let commit = build.keychain.commit(value, &key_id, switch)?;
-
-			// invert commitment
-			let commit = build.keychain.secp().commit_sum(vec![], vec![commit])?;
-
-			debug!("Building output: {}, {:?}", value, commit);
-
-			// build a proof with a rangeproof of 0 as a placeholder
-			// the test will replace this later
-			let proof = proof::create(
-				build.keychain,
-				build.builder,
-				0,
-				&key_id,
-				switch,
-				commit,
-				None,
-			)?;
-
-			// we return the output and the value is subtracted instead of added
-			Ok((
-				tx.with_output(Output::new(OutputFeatures::Plain, commit, proof)),
-				sum.sub_key_id(key_id.to_value_path(value)),
-			))
-		},
-	)
-}
-
 /// Adds an output with the provided value and key identifier from the
 /// keychain.
 pub fn output<K, B>(value: u64, key_id: Identifier) -> Box<Append<K, B>>
